Moved Array to its own submodule
