+++ startrak/native/__init__.py
@@ -2,12 +2,14 @@
 try:
 	__compiled__ = False
 	from .collections import position
+	from .collections import native_array
 	from . import classes
 	from .collections import starlist
 	from . import abstract
 except:
 	__compiled__ = True
 	from startrak.native.collections import position
+	from startrak.native.collections import native_array
 	import startrak.native.classes as classes
 	from startrak.native.collections import starlist
 	import startrak.native.abstract as abstract
@@ -16,6 +18,7 @@ except:
 Position = position.Position
 PositionLike = position.PositionLike
 PositionArray = position.PositionArray
+Array = native_array.Array
 
 Star = classes.Star
 ReferenceStar = classes.ReferenceStar
